migrated visitor

+++ src/Prometheus.Abstractions.Redesign/Language/FloatValueNode.cs
@@ -3,7 +3,7 @@ namespace Prometheus.Language
     public class FloatValueNode
         : IValueNode
     {
-        public NodeKind Kind { get; } = NodeKind.IntValue;
+        public NodeKind Kind { get; } = NodeKind.FloatValue;
         public Location Location { get; }
         public string Value { get; }
     }
